move device certificates to the settings page
