--- lib/nerves_hub_web/live/devices/edit.ex
@@ -1,47 +1,0 @@
-defmodule NervesHubWeb.Live.Devices.Edit do
-  use NervesHubWeb, :updated_live_view
-
-  alias NervesHub.Devices
-
-  def mount(%{"device_identifier" => device_identifier}, _session, socket) do
-    {:ok, device} = Devices.get_device_by_identifier(socket.assigns.org, device_identifier)
-
-    changeset = Ecto.Changeset.change(device)
-
-    socket
-    |> assign(:page_title, "Edit Device")
-    |> assign(:device, device)
-    |> assign(:form, to_form(changeset))
-    |> ok()
-  end
-
-  def handle_event("update-device", %{"device" => device_params}, socket) do
-    authorized!(:"device:update", socket.assigns.org_user)
-
-    %{org: org, product: product, device: device, user: user} = socket.assigns
-
-    message = "#{user.name} updated device #{device.identifier}"
-
-    case Devices.update_device_with_audit(device, device_params, user, message) do
-      {:ok, _device} ->
-        socket
-        |> put_flash(:info, "Device updated")
-        |> push_navigate(to: ~p"/org/#{org.name}/#{product.name}/devices")
-        |> noreply()
-
-      {:error, changeset} ->
-        socket
-        |> assign(:form, to_form(changeset))
-        |> noreply()
-    end
-  end
-
-  defp tags_to_string(%Phoenix.HTML.Form{} = form) do
-    form.data.tags
-    |> tags_to_string()
-  end
-
-  defp tags_to_string(%{tags: tags}), do: tags_to_string(tags)
-  defp tags_to_string(tags) when is_list(tags), do: Enum.join(tags, ",")
-  defp tags_to_string(tags), do: tags
-end
