--- lib/nerves_hub_web/live/devices/edit.html.heex
@@ -1,64 +1,0 @@
-<.link navigate={~p"/org/#{@org.name}/#{@product.name}/devices/#{@device.identifier}"} class="back-link">
-  Back to Device
-</.link>
-<h1>Edit Device</h1>
-
-<.form :let={f} for={@form} phx-submit="update-device">
-  <div class="form-group">
-    <div class="help-text tooltip-label help-tooltip">
-      <label for="identifier_input" class="tooltip-label">
-        ID
-      </label>
-      <span class="tooltip-info mb-1"></span>
-      <span class="tooltip-text">Device ID cannot be changed once created</span>
-    </div>
-    <%= text_input(f, :identifier, class: "form-control", id: "identifier_input", disabled: "true") %>
-    <div class="has-error"><%= error_tag(f, :identifier) %></div>
-  </div>
-
-  <div class="form-group">
-    <div class="help-text tooltip-label help-tooltip">
-      <label for="description_input" class="tooltip-label">
-        Description
-      </label>
-      <span class="tooltip-info mb-1"></span>
-      <span class="tooltip-text">Description for this device</span>
-    </div>
-    <%= text_input(f, :description, class: "form-control", id: "description_input") %>
-    <div class="has-error"><%= error_tag(f, :description) %></div>
-  </div>
-
-  <div class="form-group">
-    <div class="help-text tooltip-label help-tooltip">
-      <label for="tags_input" class="tooltip-label">
-        Tags
-      </label>
-      <span class="tooltip-info mb-1"></span>
-      <span class="tooltip-text">Tags are used by deployments to target a device. A device must have matching tag(s) for the deployment to update it</span>
-    </div>
-    <%= text_input(f, :tags, class: "form-control", id: "tags_input", value: tags_to_string(@form)) %>
-    <small class="form-text text-muted mt-1">
-      Every device is added to the 'all' tag by default. You can have multiple tags separated by commas.
-    </small>
-    <div class="has-error"><%= error_tag(f, :tags) %></div>
-  </div>
-
-  <div class="form-group">
-    <div class="help-text tooltip-label help-tooltip">
-      <label for="connecting_code" class="tooltip-label">
-        First Connect Code
-      </label>
-      <span class="tooltip-info mb-1"></span>
-      <span class="tooltip-text">Run this code when the device first connects to the console.</span>
-    </div>
-    <%= textarea(f, :connecting_code, rows: 6, class: "form-control", id: "connecting_code") %>
-    <small class="form-text text-muted mt-1">
-      Make sure this is valid Elixir and will not crash the device.
-    </small>
-    <div class="has-error"><%= error_tag(f, :connecting_code) %></div>
-  </div>
-
-  <div class="button-submit-wrapper">
-    <%= submit("Save Changes", class: "btn btn-primary") %>
-  </div>
-</.form>
