+++ lib/nerves_hub_web/live/devices/settings.html.heex
@@ -1,0 +1,164 @@
+<.link navigate={~p"/org/#{@org.name}/#{@product.name}/devices/#{@device.identifier}"} class="back-link">
+  Back to Device
+</.link>
+<h1>Device Settings</h1>
+
+<.form :let={f} for={@form} phx-submit="update-device">
+  <div class="form-group">
+    <div class="help-text tooltip-label help-tooltip">
+      <label for="identifier_input" class="tooltip-label">
+        ID
+      </label>
+      <span class="tooltip-info mb-1"></span>
+      <span class="tooltip-text">Device ID cannot be changed once created</span>
+    </div>
+    <%= text_input(f, :identifier, class: "form-control", id: "identifier_input", disabled: "true") %>
+    <div class="has-error"><%= error_tag(f, :identifier) %></div>
+  </div>
+
+  <div class="form-group">
+    <div class="help-text tooltip-label help-tooltip">
+      <label for="description_input" class="tooltip-label">
+        Description
+      </label>
+      <span class="tooltip-info mb-1"></span>
+      <span class="tooltip-text">Description for this device</span>
+    </div>
+    <%= text_input(f, :description, class: "form-control", id: "description_input") %>
+    <div class="has-error"><%= error_tag(f, :description) %></div>
+  </div>
+
+  <div class="form-group">
+    <div class="help-text tooltip-label help-tooltip">
+      <label for="tags_input" class="tooltip-label">
+        Tags
+      </label>
+      <span class="tooltip-info mb-1"></span>
+      <span class="tooltip-text">Tags are used by deployments to target a device. A device must have matching tag(s) for the deployment to update it</span>
+    </div>
+    <%= text_input(f, :tags, class: "form-control", id: "tags_input", value: tags_to_string(@form)) %>
+    <small class="form-text text-muted mt-1">
+      Every device is added to the 'all' tag by default. You can have multiple tags separated by commas.
+    </small>
+    <div class="has-error"><%= error_tag(f, :tags) %></div>
+  </div>
+
+  <div class="form-group">
+    <div class="help-text tooltip-label help-tooltip">
+      <label for="connecting_code" class="tooltip-label">
+        First Connect Code
+      </label>
+      <span class="tooltip-info mb-1"></span>
+      <span class="tooltip-text">Run this code when the device first connects to the console.</span>
+    </div>
+    <%= textarea(f, :connecting_code, rows: 6, class: "form-control", id: "connecting_code") %>
+    <small class="form-text text-muted mt-1">
+      Make sure this is valid Elixir and will not crash the device.
+    </small>
+    <div class="has-error"><%= error_tag(f, :connecting_code) %></div>
+  </div>
+
+  <div class="button-submit-wrapper">
+    <%= submit("Save Changes", class: "btn btn-primary") %>
+  </div>
+</.form>
+
+<div class="border-bottom border-dark mt-6 mb-4"></div>
+
+<div class="device-header-group">
+  <h3 class="mb-2">Certificates</h3>
+  <%= if @toggle_upload do %>
+    <button class="btn btn-primary" type="button" phx-click="toggle-upload" phx-value-toggle={to_string(@toggle_upload)}>
+      <span class="action-text">Cancel Upload</span>
+    </button>
+  <% else %>
+    <button class="btn btn-outline-light btn-action" type="button" phx-click="toggle-upload" phx-value-toggle={to_string(@toggle_upload)}>
+      <div class="button-icon add"></div>
+    </button>
+  <% end %>
+</div>
+
+<div class={unless @toggle_upload, do: "hidden"}>
+  <form phx-change="validate-cert" phx-drop-target={@uploads.certificate.ref}>
+    <div class="form-group import-pem">
+      <label for={@uploads.certificate.ref}>Drag or Click to upload</label>
+      <.live_file_input upload={@uploads.certificate} class="hidden" />
+    </div>
+  </form>
+  <br />
+</div>
+
+<%= if Enum.any?(@device.device_certificates, & &1.org_id != @device.org_id) do %>
+  <div class="alert alert-warning">
+    <div class="content-container">
+      <center>Device was previously moved and has certificates associated with a Signer CA from another organization</center>
+      <br />
+      <center>In most cases, this should be okay (or was even intentional), but automatic device certificate creation</center>
+      <center>will not work until the Signer CA is also moved the the same organization as this device.</center>
+      <br />
+      <center>If the intention is leave the Signer CA in the other organization, then this will continue to work</center>
+      <center>so long as certificates are managed manually for this device and other metadata, such as the product</center>
+      <center>and firmware signing keys appropiately match the existing firmware as well</center>
+    </div>
+  </div>
+<% end %>
+<table class="table table-sm table-hover">
+  <thead>
+    <tr>
+      <th>Serial</th>
+      <th>Last used</th>
+      <th>Not before</th>
+      <th>Not after</th>
+    </tr>
+  </thead>
+  <tbody>
+    <%= for cert <- @device.device_certificates do %>
+      <% mismatch? = cert.org_id != @device.org_id %>
+      <tr class="item" style={if mismatch?, do: "border: 2px dashed var(--warning)"}>
+        <td>
+          <div class="mobile-label help-text">Serial</div>
+          <code class="color-white wb-ba ff-m"><%= Utils.format_serial(cert.serial) %></code>
+        </td>
+        <td title={cert.last_used}>
+          <div class="mobile-label help-text">Last used</div>
+          <%= if !is_nil(cert.last_used) do %>
+            <%= DateTimeFormat.from_now(cert.last_used) %>
+          <% else %>
+            <span class="text-muted">Never</span>
+          <% end %>
+        </td>
+        <td>
+          <div class="mobile-label help-text">Not before</div>
+          <div class="date-time"><%= cert.not_before %></div>
+        </td>
+        <td>
+          <div class="mobile-label help-text">Not after</div>
+          <div class="date-time"><%= cert.not_after %></div>
+        </td>
+
+        <td class="actions">
+          <div class="mobile-label help-text">Actions</div>
+          <div class="dropdown options">
+            <a class="dropdown-toggle options" href="#" id={"cert-#{cert.id}"} data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+              <div class="mobile-label pr-2">Open</div>
+              <img src="/images/icons/more.svg" alt="options" />
+            </a>
+            <div class="dropdown-menu dropdown-menu-right">
+              <%= if mismatch? do %>
+                <%= link("Organization", class: "dropdown-item", aria_label: "Originating Organization", to: ~p"/org/#{Repo.preload(cert, :org).org.name}") %>
+              <% end %>
+              <%= if cert.der do %>
+                <%= link("Download",
+                  class: "dropdown-item",
+                  aria_label: "Download Device Certificate",
+                  to: Routes.device_path(@socket, :download_certificate, @org.name, @product.name, @device.identifier, cert.serial)
+                ) %>
+              <% end %>
+              <a class="dropdown-item" phx-click="delete-certificate" phx-value-serial={cert.serial} data-confirm={"Delete #{Utils.format_serial(cert.serial)}?"}>Delete</a>
+            </div>
+          </div>
+        </td>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
