+++ lib/nerves_hub_web/live/devices/show.ex
@@ -4,7 +4,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
   require Logger
 
   alias NervesHub.AuditLogs
-  alias NervesHub.Certificate
   alias NervesHub.Deployments
   alias NervesHub.Devices
   alias NervesHub.Devices.UpdatePayload
@@ -13,8 +12,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
   alias NervesHub.Tracker
 
   alias NervesHubWeb.Components.DeviceHeader
-  alias NervesHubWeb.Components.Utils
-  alias NervesHubWeb.LayoutView.DateTimeFormat
 
   alias Phoenix.Socket.Broadcast
 
@@ -33,57 +30,10 @@ defmodule NervesHubWeb.Live.Devices.Show do
       |> assign(:status, Tracker.status(device))
       |> assign(:deployment, device.deployment)
       |> assign(:page_title, device.identifier)
-      |> assign(:toggle_upload, false)
       |> assign(:results, [])
       |> assign(:deployments, Deployments.alternate_deployments(device))
       |> assign(:firmwares, Firmwares.get_firmware_for_device(device))
-      |> allow_upload(:certificate,
-        accept: :any,
-        auto_upload: true,
-        max_entries: 1,
-        progress: &handle_progress/3
-      )
       |> audit_log_assigns(1)
-
-    {:ok, socket}
-  end
-
-  def handle_progress(:certificate, %{done?: true} = entry, socket) do
-    socket =
-      socket
-      |> clear_flash(:info)
-      |> clear_flash(:error)
-      |> consume_uploaded_entry(entry, &import_cert(socket, &1.path))
-
-    {:noreply, socket}
-  end
-
-  def handle_progress(:certificate, _entry, socket), do: {:noreply, socket}
-
-  defp import_cert(%{assigns: %{device: device}} = socket, path) do
-    socket =
-      with {:ok, pem_or_der} <- File.read(path),
-           {:ok, otp_cert} <- Certificate.from_pem_or_der(pem_or_der),
-           {:ok, db_cert} <- Devices.create_device_certificate(device, otp_cert) do
-        updated = update_in(device.device_certificates, &[db_cert | &1])
-
-        assign(socket, :device, updated)
-        |> put_flash(:info, "Certificate Upload Successful")
-      else
-        {:error, :malformed} ->
-          put_flash(socket, :error, "Incorrect filetype or malformed certificate")
-
-        {:error, %Ecto.Changeset{errors: errors}} ->
-          formatted =
-            Enum.map_join(errors, "\n", fn {field, {msg, _}} ->
-              ["* ", to_string(field), " ", msg]
-            end)
-
-          put_flash(socket, :error, IO.iodata_to_binary(["Failed to save:\n", formatted]))
-
-        err ->
-          put_flash(socket, :error, "Unknown file error - #{inspect(err)}")
-      end
 
     {:ok, socket}
   end
@@ -172,23 +122,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     {:noreply, assign(socket, :device, Repo.preload(updated_device, [:device_certificates]))}
   end
 
-  def handle_event(
-        "delete-certificate",
-        %{"serial" => serial},
-        %{assigns: %{device: device}} = socket
-      ) do
-    certs = device.device_certificates
-
-    with db_cert <- Enum.find(certs, &(&1.serial == serial)),
-         {:ok, _db_cert} <- Devices.delete_device_certificate(db_cert),
-         updated_certs = Enum.reject(certs, &(&1.serial == serial)) do
-      {:noreply, assign(socket, device: %{device | device_certificates: updated_certs})}
-    else
-      _ ->
-        {:noreply, put_flash(socket, :error, "Failed to delete certificate #{serial}")}
-    end
-  end
-
   def handle_event("restore", _, socket) do
     authorized!(:"device:restore", socket.assigns.org_user)
 
@@ -218,15 +151,9 @@ defmodule NervesHubWeb.Live.Devices.Show do
     {:noreply, assign(socket, :device, device)}
   end
 
-  def handle_event("toggle-upload", %{"toggle" => toggle}, socket) do
-    {:noreply, assign(socket, :toggle_upload, toggle != "true")}
-  end
-
   def handle_event("clear-flash-" <> key_str, _, socket) do
     {:noreply, clear_flash(socket, String.to_existing_atom(key_str))}
   end
-
-  def handle_event("validate-cert", _, socket), do: {:noreply, socket}
 
   def handle_event("push-update", %{"uuid" => uuid}, socket) do
     authorized!(:"device:push-update", socket.assigns.org_user)
