+++ test/nerves_hub_web/live/devices/settings_test.exs
@@ -1,13 +1,13 @@
-defmodule NervesHubWeb.Live.Devices.EditTest do
+defmodule NervesHubWeb.Live.Devices.SettingsTest do
   use NervesHubWeb.ConnCase.Browser, async: true
 
   alias NervesHub.Devices
 
-  describe "edit device" do
+  describe "device settings" do
     test "can change tags", %{conn: conn, org: org, product: product, device: device} do
       conn
-      |> visit("/org/#{org.name}/#{product.name}/devices/#{device.identifier}/edit")
-      |> assert_has("h1", text: "Edit Device")
+      |> visit("/org/#{org.name}/#{product.name}/devices/#{device.identifier}/settings")
+      |> assert_has("h1", text: "Device Settings")
       |> fill_in("Tags", with: "josh, lars")
       |> click_button("Save Changes")
       |> assert_path("/org/#{org.name}/#{product.name}/devices")
@@ -18,8 +18,8 @@ defmodule NervesHubWeb.Live.Devices.EditTest do
 
     test "can add 'first connect code'", %{conn: conn, org: org, product: product, device: device} do
       conn
-      |> visit("/org/#{org.name}/#{product.name}/devices/#{device.identifier}/edit")
-      |> assert_has("h1", text: "Edit Device")
+      |> visit("/org/#{org.name}/#{product.name}/devices/#{device.identifier}/settings")
+      |> assert_has("h1", text: "Device Settings")
       |> fill_in("First Connect Code", with: "dbg(\"boo\")")
       |> click_button("Save Changes")
       |> assert_path("/org/#{org.name}/#{product.name}/devices")
